Move base.config into cpy.
This is to make cpy less dependent on the rest of pysph.
