+++ pysph/base/tests/test_nnps.py
@@ -11,7 +11,7 @@ from numpy import random
 from pysph.base.point import IntPoint, Point
 from pysph.base.utils import get_particle_array
 from pysph.base import nnps
-from pysph.base.config import get_config
+from pysph.cpy.config import get_config
 
 # Carrays from PyZoltan
 from pyzoltan.core.carray import UIntArray, IntArray
