+++ pysph/cpy/api.py
@@ -1,5 +1,6 @@
 from .ast_utils import (get_symbols, get_assigned,
                         get_aug_assign_symbols, has_return, has_node)
+from .config import get_config, set_config, Config
 from .cython_generator import (
     CythonGenerator, KnownType, declare, get_func_definition
 )
