+++ pysph/cpy/tests/test_translator.py
@@ -2,7 +2,7 @@ from textwrap import dedent
 import pytest
 import numpy as np
 
-from pysph.base.config import get_config
+from ..config import get_config
 from ..translator import (
     CConverter, CodeGenerationError, CStructHelper, KnownType,
     OpenCLConverter, py2c
