+++ pysph/sph/acceleration_eval_opencl_helper.py
@@ -84,11 +84,11 @@ import pyopencl as cl
 import pyopencl.array  # noqa: 401
 import pyopencl.tools  # noqa: 401
 
-from pysph.base.config import get_config
 from pysph.base.utils import is_overloaded_method
-from pysph.cpy.ext_module import get_platform_dir
 from pysph.base.opencl import (profile_kernel, get_context, get_queue,
                                DeviceHelper)
+from pysph.cpy.ext_module import get_platform_dir
+from pysph.cpy.config import get_config
 from pysph.cpy.translator import (CStructHelper, OpenCLConverter,
                                   ocl_detect_type)
