+++ pysph/sph/tests/test_acceleration_eval.py
@@ -10,8 +10,8 @@ import pytest
 import numpy as np
 
 # Local imports.
-from pysph.base.config import get_config
 from pysph.base.utils import get_particle_array
+from pysph.cpy.config import get_config
 from pysph.cpy.api import declare
 from pysph.sph.equation import Equation, Group
 from pysph.sph.acceleration_eval import (
